Final commit with README, working tests, some bugs fixed on the frontend, security in the form inputs for the transaction

+++ db/migrate/20180907085023_create_accounts.rb
@@ -1,7 +1,7 @@
 class CreateAccounts < ActiveRecord::Migration[5.1]
   def change
     create_table :accounts do |t|
-      t.decimal :balance, default: 0
+      t.float :balance, default: 0
       t.integer :user_id
       t.belongs_to :user, index: true
